D2Tree with initialization
Given a minimum height, the necessary amount of nodes is manipulated to fill a D2Tree of that size. If there are any extra nodes, they are added like they used to.

A D2Tree of height h, has 2^h - 1 nodes in its PBT.
It also has h nodes in each bucket (when balanced) and 2^(h -1) buckets.

Therefore, nodespace has to be greater than or equal to (2^h - 1) + h * 2^(h - 1).